Functioning basic cpp function o split latitude and longitude when together in a string
Tests added for the split function

There are still issues when testing scrub and scrup_cpp because tests do not use UTF-8 encoding.
